refactor: promise系のutilを分けた
