+++ src/speaker/ttsControllerSpeaker.ts
@@ -6,8 +6,8 @@ import { Readable } from "stream";
 
 import { Session } from "../session";
 import { AdditionalVoiceParam, SpeechText, VoiceParam } from "../types";
+import { wait } from "../util/promiseUtil";
 import { Result, success } from "../util/result";
-import { wait } from "../util/util";
 import { SIOAudioRecorder } from "./socketIOAudioRecorder";
 import { Speaker, SpeakerState } from "./speaker";
 import { ttsControllerOccupier } from "./ttsControllerOccupier";
