+++ src/speaker/voiceProvider.ts
@@ -5,8 +5,8 @@ import { getLogger } from "log4js";
 import { Session } from "../session";
 import { SpeakerOption, SpeechText } from "../types";
 import { Deferred } from "../util/deferred";
+import { allSerial } from "../util/promiseUtil";
 import { failure, Result } from "../util/result";
-import { allSerial } from "../util/util";
 import { Speaker } from "./speaker";
 import { TtsControllerSpeaker, TtsSpeakerBuildOption } from "./ttsControllerSpeaker";
 import { DaemonSpeakerBuildOption, VoiceroidDaemonSpeaker } from "./voiceroidDaemonSpeaker";
